Update patch set 2
Patch Set 2:

A Buck contributor already contacted us about SuperDevMode embedding, but I convinced him to use the existing API [1].

The Buck patch [2] is stalled but I think that's due to Jetty version issues.

[1] https://gwt-review.googlesource.com/#/c/8772/
[2] https://gerrit-review.googlesource.com/#/c/59354/

I'm speculating on monitoring but it seems safe to assume someone will want to know about the basic changes to a job like when it starts and stops compiling. Maybe not the more specialized fields we will need, but we can document them as experimental.

Patch-set: 2