Update patch set 1
Patch Set 1:

Goktug was suggesting that we pass these flags in the "gwt.args" system property to be consistent with GWTTestCase.

However, we looked at the code, and GWTTestCase passes the flags from "gwt.args" to DevMode. I suspect that if we run GWTTestCase using -Dgwt.args='-superdevmode' then GWTTestCase will launch DevMode and this will launch Super Dev Mode. I doubt this works today, but we should probably make it work.

So, perhaps we should just add the -XjsInteropMode and -XcompilePerFile flags to DevMode's flags and copy them into Super Dev Mode's flags. Then when we make GWTTestCase work properly, we we be able to pass something like -Dgwt.args='-superdevmode -XcompilePerFile=true' and it will work properly.

Patch-set: 1