Update patch set 1
Patch Set 1:

To avoid the circular dependency you might have to move Options.CompilerPerFileFlag next to the other flag classes, with a similar structure as ArgHandlerJsInteropMode which is already in the right place.

(This is a little weird because classic DevMode doesn't actually support these two flags. They would only be used for Super Dev Mode. But it does avoid having flags within flags so I still think it's better.)

Patch-set: 1