Update patch set 5
Patch Set 5:

Thanks for fixing the bug in the alert dialog. It looks okay but I noticed something else we might want to fix:

I originally intended computeScriptBase.js to be used only when Super Dev Mode is on. The idea was to search for the script tag that the original nocache.js file just created in order to redirect to Super Dev Mode. Since Super Dev Mode created the script tag, it doesn't need to handle other things like meta tags that the full version of computeScriptBase.js [1] must handle.

Since the stub is running before Super Dev Mode starts up, I think it should be using something like the full version of computeScriptBase. I'm not sure if it's better to include it or inline in into stub.nocache.js.

[1] dev/core/src/com/google/gwt/core/ext/linker/impl/computeScriptBase.js

Patch-set: 5
Label: Verified=0