Makes Generator output invalidation more accurate.
Previously all output for any Generator that was being rerun as a result
of modified resources or types was assumed stale.

For Generators with large amounts of output (Gin, GWTRPC) this was very
costly. This change stops making that assumption and instead watches
the types created by Generators, flags the ones that have changed as
modified and recalculates the stale types list.

Previously UnifyAst assumed that the stale types list did not change, so
extra care had to be taken there.

Change-Id: I5f296eacb30bb6abc22d9daeccb87423c1766736