Update patch set 7
Patch Set 7:

> 
 > Since the stub is running before Super Dev Mode starts up, I think
 > it should be using something like the full version of
 > computeScriptBase. I'm not sure if it's better to include it or
 > inline in into stub.nocache.js.
 > 

I think when I implemented the stub.nocache in this way I took a look to the full computeStringBase but I dont remember why I discarded it.

I think replace computeScriptBase in the stub js should be discussed in a different patch, so lets continue fixing these couple of things here if you are ok with them.

Patch-set: 7
Label: Verified=0