Update patch set 2
Patch Set 2:

We could put it in, but if we don't test it regularly then I'm concerned that it will stop working and we won't notice. It seems like to test this code, we need to have jetty-annotations.jar available somehow (but it doesn't have to be in gwt-dev.jar) and a test that uses Servlet 3.0 annotations.

(But on the other hand there's a lot of other untested code so I won't hold this up if other folks are okay with it.)

Patch-set: 2