Update patch set 1
Patch Set 1:

So, an update on this: Daniel is going to rename the -XcompilePerFile flag to -incremental and we're going to make incremental compiles the default soon. See [1]. I think we still want to be able to opt out of -incremental though (for 2.7), so it makes sense to support it here to turn it off.

[1] https://gwt-review.googlesource.com/#/c/9211/

Patch-set: 1
Label: Code-Review=0