Update patch set 1
Patch Set 1:

Not quite ready for review yet. I'm not happy with the way Defender methods are implemented. I'd like to move it to GwtAstBuilder but I need to figure out how. I originally tried implementing it there, but ended up with duplicate classes in the AST (JClassType's not 'interned', after UnifyAst, two copies would exist in the AST with different memory locations)

Doing it post-unification solved the problem, but it is unsatisfying.

Patch-set: 1
Label: Verified=0