Update patch set 1
Patch Set 1:

CI server is running java7, I guess some time we have to upgrade it if we want to make tests run. I don't know if we should have flags or checks for running java8 tests if the platform is java8, or if it's worth to maintain builds with both versions for a while.

Patch-set: 1
Label: Verified=0