Update patch set 3
Patch Set 3:

(10 comments)

I'll send the updated patch later as it is already late today.

One important consideration in the review is also making sure everything is behaving reasonable without the checks that are not marked as critical.

For example, removing the check in ArrayList#get is ok because it will return null and that will probably eventually cause NPE. However if we remove the check in LinkedList, that will potentially cause infinite loop, which is not OK.

Patch-set: 3