Recompile current module on reload if started through dev mode.
If super dev mode is started from the regular dev mode target,
we synthesize a nocache.js file on the fly.

This patch adds a file called $MODULE_NAME.recompile.nocache.js
that if loaded will trigger a recompile of the current module
and start the newly compilated app in the current page.
This patch changes the synthesized nocache.js to just load the
recompile.nocache.js file and thus trigger a recompile on every
reload.

Also changes super dev mode server side to always emit a
recompile.nocache.js for every compile.

Change-Id: I6b3fdc441c2a73f9f627ca7a766023db0ccf68b8