Unignores CSS changes in .ui.xml files in incremental compiles.
Previously changes to css inside of style tags nested inside of .ui.xml
template files were not being respected during incremental compiles.

The problem was that UIBinderGenerator's .java output was being
invalidated and reprocessed but the changes that reflect css inside of
style tags actually came out of InlineClientBundleGenerator (even though
it never reads the .ui.xml template file).

Basically there is a back door mechanism in user land Generator code
where Generator A can create a synthetic resource and pass it along as
input to Generator B, whose behavior is now modified.

Since GWT is not a build system and does not model this kind of
arbitrary user land Generator dependency, and since there is no intention
to make this behavior ok in the future, it was decided to change
UIBinderGenerator to change its .java output when *any* change occurs to
it's template file. This ensure that it's InlineClientBundleGenerator
dependency is rerun.

In a saner world, Generator execution would be managed by a build system
(like Buck, Gradle, etc) that models both real and synthetic inputs and
outputs of Generators, cache freshness and minimal reinvocation. Maybe
we can work within such a system in the future.

Change-Id: I4737b8cfbe7fc2fe9f2ec44e5d331522ee3f7eab