Update patch set 1
Patch Set 1:

Thanks for fixing this up.

Yes we do want to run most tests in incremental mode, but not all. Since CodeSplitting and Optimization are inherently global steps, they can't be done incrementally.

We actually have test coverage of all of the compatible types of tests running incrementally internally, but not externally in the Ant files.

I think we're kind of waiting for something like Buck or Gradle to get mature enough that we can use it for both internal and external build system (and get rid of Ant).

Patch-set: 1