Update patch set 1
Patch Set 1:

(6 comments)

I think we should continue to call dev_mode_on.js to display the dialog and do the actual compile. Maybe that code could be improved, but it's known to work on IE8 and using the same compile dialog is more consistent.

Regarding the progress dialog, the dots are a placeholder and the plan is to provide actual progress about what the compiler is actually up to. Of course when the compile is instant, it doesn't matter, but on the first compile I think it's still worth displaying progress info.

Patch-set: 1
Label: Verified=0