Simplifies the way linkers add code at the beginning of the output.
Linkers used a complex process to keep the output JS and the corresponding
sourcemaps in sync by adding a synthetic artifact that represented the
editions of the source.

This patch simplifies the process by only letting the linker add code the
first line (line breaks are stripped from the module prefix) thus making the
sourcemap artifact be valid after edition.

Change-Id: If4e033ef92a95172f739ef7dbb88de49b2369c87