Makes SDM MinimalRebuildCache updates transactional.
When SDM invokes incremental compiles it takes care to pass the proper
MinimalRebuildCache instance (for the current binding properties) to
the Compiler, and the Compiler updates the cache in place.

But if the compile failed it was possible for the cache to be left in
an inconsistent state (for example having already had some entries
removed but not yet had them rebuilt).

This change teaches SDM to always pass the Compiler a copy of the
current known-good MinimalRebuildCache and only replace its copy with
a modified one after a successful compile. This way whenever a compile
starts it will always be starting with a known good MinimalRebuildCache
instance.

Change-Id: I1d536c01746f6ede846cf0ddc519267f3809b6df