Update patch set 6
Patch Set 6: Code-Review-1

> about 1:
 > Bookmarklets should go away since compiles are now fast enough that
 > you can always just recompile.

I continue thinking that recompiling always is a bad idea. It should be launched automatically when the first compilation was not made and let the user decide when to compile, like currently it is.

Right now incremental is slower than noincremental when compiling Vaadin widgetsets, first compilation (18.1 vs 11.2 sec) , and successive ones (20.4 vs 6.7ms).

In vaadin projects we don't want to launch recompile always, we change UI in server so we reload the page but want to reuse last compiled widgetset.  Sometimes these changes modify some share code that we don't want to populate still to the widgetset. So the ideal situation is let the developer decide whether to compile.

In embedded apps and web-views compiling through reloading the app is difficult, while a button like we have currently is more convenient.

 > 
 > about 2:
 > I do not see the point in dots as progress, its extra complexity
 > and does not gie any extra information, we could just have a timer
 > adding the dots.

Progress is fundamental to know that server continues working in large projects, and it's the only way to detect something went wrong like memory errors or server stuck. Fake dots or animation is not an option.
Also we should ask whether the server is available before send the compile json.

Finally, I don't see the advantage of rewriting completely new code in recompile.nocache vs the stub.nocache and dev_mode_on which we have been tested during months, especially when this new code is breaking some features and changing the behaviour. I'm ok if it goes in 3.0 but I think is a bit hasty to have it in 2.7.

Patch-set: 6
Label: Code-Review=-1