Update patch set 6
Patch Set 6:

Hi Manolo,

I already highlighted to the vaadin team why there compile is slow and they are working on fixing that.

I feel very strongly that this is the right approach going forward since it is way simpler than before. It works like the old dev mode and people do not have to deal with setting up bookmarklets etc.

Our big projects (which are way bigger than vaadin), change detection is always under a second and they manage to compile in less than 10s (most even faster), so this is something in your setup. Most outside projects are not that large and will be mostly under 5s (most under 1s), so we really do not need a progress indicator, since the developer can see the progress in the console anyway.

The old code in stub.nocache.js was really hacky, full of global variables and closures. Rewriting this is long overdue anyway.

Again I feel strongly that this is the right feature for GWT developers going forward. It reduces complexity and makes SDM as easy to use as the old dev mode. WIth the work that brandon is doing on eclipse, we are getting a really good story.

If your compile is too slow in vaadin (after you guys implemented my suggestions) I am fine to look at it again, but I am fairly certain that compile is not an issue going forward.

Patch-set: 6