Update patch set 6
Patch Set 6:

Actually I think the key point is the ability to skip the compile if there are no changes. This happens much faster (under 500ms). So, if you are not editing code then there shouldn't be a delay.

I think it makes sense for recompile-on-reload to be the default for DevMode -superdevmode since it's easy to use and matches how DevMode worked before. However, perhaps we should have a way to turn it off and use the bookmarklet instead, for people who prefer fine control over convenience.

Manolo, could you explain why you'd want to use server code that doesn't match client code? I know this happens sometimes but I didn't think it was something you'd want to do deliberately.

Patch-set: 6