Update patch set 8
Patch Set 8:

Daniel and I talked about it. In the cases where the browser reload button is somehow not available, we think it should be easy to add an HTML button to the page that just calls window.location.reload(). It's likely that you'll want a reload button in any case. So we don't think DevMode -superdevmode needs to provide this.

Patch-set: 8
Label: Verified=0