Update patch set 9
Patch Set 9:

> Manolo, could you explain why you'd want to use server code that doesn't match client code? 

One case is when making changes in a Vaadin ComponentState class. If I change the default flag value below to false it’s not necessary to modify the client side. There is more shared stuff which is not used in client unless you are developing offline apps, so makes not sense to recompile if the app is for online apps.

  public class MyComponentState  extends AbstractFieldState {
   // changing this does not makes necessary compile client
    public boolean flag = true; 
 }


The solution having a flag to run the superdevmode without recompiling on reload has to be coexist with someway to force compilation like bookmarklet or just do the compilation the first time if the flag is set. A simpler solution could be not have that flag but prevent to recompile when there is some parameter like index.html?recompile=false

Since classic dev mode works in the same way I agree that it’s coherent that by default reloading recompiles.


> It's likely that you'll want a reload button in any case. So we don't think DevMode -superdevmode needs to provide this.

But actually recompile.nocache provides it and it’s shown when compilation fails. I think it could be shown/hidden using CSS. In fact I think almost UI stuff in recompile.nocache should be styled using css like we were doing in stub.nocache instead of inlining them, in this way, it’s easy to brand it in the host page.

Patch-set: 9