Update patch set 9
Patch Set 9: Code-Review+2

In the interest of getting things done quickly, I think the way forward is to commit this patch and have followup patches to add back some of the functionality it drops.

In particular, I will take this on:
- we need to write public resources to the DevMode directory
- get the bookmarklets working again

Daniel is looking into an issue with the wrong permutation being selected after switching user agent.

I'm open to adding back some kind of minimal in-page UI. In particular, it would be nice to have some kind of indicator showing that Super Dev Mode is on.

Patch-set: 9
Label: Code-Review=+2