Update patch set 1
Patch Set 1:

> (2 comments)
 > 
 > As I explained in a comment on issue 8898, I believe this behavior
 > is by-design.

I think user does not need to know the internals of how a null string is maintained internally as a reference to the json null object, so it makes sense to me to handle it in these methods.

 > 
 > If we change the behavior for JsonObject#put(String,String), we
 > should also change the behavior for JsonArray#set(int,String)

This patch changes both classes. Are you pointing to a different class?

Patch-set: 1