Update patch set 2
Patch Set 2:

Oh my! Just noticed this would break the gwt-maven-plugin, and fixing it  is going to be a PITA given the state of Maven APIs :-(

That doesn't mean we should hold this patch though, there's an easy workaround: add a direct dependency on gwt-dev in your project so you get transitive dependencies (something the plugin doesn't honor when constructing forked processes' classpath).

Patch-set: 2