Update patch set 2
Patch Set 2:

> @Thomas I'm wondering if we could do the same with jetty, maybe in
 > this patch or a different one.

The problem with Jetty (and HTMLUnit FWIW) is that AFAICT we rely on a specific version, so unbundling it would make it easy to break DevMode, SuperDevMode or JUnitShell by using an incompatible version (not to mention that HTMLUnit depends on Jetty).

ASM and Guava on the other hand have (rather) clear compatibility policies, so using a newer Guava or ASM should work. This is not the case with Jetty.

Maybe we could specify a version range though for those, to make it clear that GWT is not compatible with, say, Jetty 9, but I have no experience with version ranges and I've been told to avoid them if possible (and we would have to test with at least the latest version of Jetty 8 to determine which upperbound to set for the range).

Don't get me wrong: I'm OK with unbundling as much as possible, but this is not an easy task, so let's do it step by step (and honestly I'd rather do it while modularizing GWT, with a new –maintainable– build system).

#yakshaving

Patch-set: 2