Update patch set 5
Patch Set 5:

> This change removes properties gwt.logging.firebugHandler and
 > gwt.logging.popupHandler, this properties are sometimes (ex.
 > PhoneGap and Errai) used just to be set to DISABLED, does not worth
 > leaves this properties defined and deprecated so this change will
 > not break this libraries?

Unfortunately there is no deprecation concept for properties. So if I leave them now in this release, then nobody will notice and remove so we will be in the same state in the next release when we remove them.
Slight advantage would be if a library tries to support both 2.6 and 2.7 at the same time but not sure if that's an important use case as 2.7 is not a big breaking release.

Patch-set: 5
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0