Super Dev Mode: add -launcherDir flag
This provides a way to start a GWT application using Super Dev Mode
without having done a full compile of the GWT application ahead of
time.

The -launcherDir flag is optional and points to an output directory.
The code server will create a subdirectory for each module. On
startup and after each recompile, it will write out a nocache.js
file and any public resources in the module. (Note that we don't delete
files except for a nocache.js.gz file if it exists.)

For example, -launcherDir can be pointed to the expanded war
directory of a Java frontend server.

An upcoming patch will switch devmode's -superDevMode flag to use this
mechanism instead.

Change-Id: Iced1fb38d85d225c92e2d02b30e53675e1a18bfb