Fixes a bug in JsStaticEval.
JsStatic eval used to change the order in which binary operations
are performed in an attempt to reduce the number of parenthesis
required to represent expressions.

The method to perform this transformations is buggy and poorly
written. This patch removes this method (correctness is
much more important than codesize). In particular completely
ignores the associativity of some of the operators it is moving
around.

If it turns out to be a significant increase in code size, the
transformation needs to be rewritten.

JsStaticEvalTest preserves the tests in commented form for the
desired outcomes of such a pass.

Change-Id: I66f0a7c770ac376f530003acaf171e3a99815160