Explicitly bound ie9 for FocusImpl and LayoutImpl
This commit does not change the behavior of GWT, as when unbound ie9
currently falls back to these implementations, despite the 'do not fall
back' comment and <none> tag, since this rule is still the closest
match.

This commit only avoid warning log messages about the fallback, future
commits might explicitly rebind ie9 to a standards-based implementation
instead.

Reverts the intention of 040b3e43, but is aimed at keeping the same
behavior.

Bug: issue 8848, issue 8383
Change-Id: I03a4e1083438510a9e60c4ee613d4902811df028