Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

the change looks good as a way of preventing users from starting to opt into the library compilation pathway that is still just experimental. but is this a change that we want to commit in the main line or just in the 2.7 branch?

Patch-set: 1
Reviewer: Gerrit User 5410 <5410@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+2