Update patch set 2
Patch Set 2:

> 
 > The problem with Jetty (and HTMLUnit FWIW) is that AFAICT we rely
 > on a specific version, so unbundling it would make it easy to break
 > DevMode, SuperDevMode or JUnitShell by using an incompatible
 > version (not to mention that HTMLUnit depends on Jetty).
 > 

My idea is simply remove jetty from the gwt-dev bundle deployed in maven repos, no from the gwt-dev bundled with the sdk.zip, but make gwt-dev.pom depends on the current jetty version.

I think this does not break anything, but gives some advantages to the user who knows exactly the dependency graph of its project, and could do things under his risk like define exclusions and try or mix different versions of jetty  jars, etc. 

So the main goal is to have a smaller gwt-dev.jar, and to know exactly that gwt-dev depends on jetty 8.1.12.v20130726 when you use a  build system with dependency management.

A range of compatible jetty could be Ok, but I would do/discuss in a second step., after we can experiment with the split version of gwt-dev and different versions of jetty.

Patch-set: 2