Update patch set 2
Patch Set 2:

> Oh my! Just noticed this would break the gwt-maven-plugin, and
 > fixing it  is going to be a PITA given the state of Maven APIs :-(
 > 
 
I think adding the gwt-dev dependency with a provided scope is the natural way to configure a maven project, I always do seeing the nag message from the plugin complaining about that.
I would get rid of gwt-dev dependency checks in the plugin, I think this is something inherited from the years where the plugin was taking care of downloading the sdk for a concrete SO and unziping it, etc.

Patch-set: 2