Update patch set 2
Patch Set 2:

> 
 > Well, it could break projects that:
 > * have Jetty as a dependency

I don't understand this problem, if a project depends on say jetty-9 and we say that gwt-dev depends on jetty-8, is the same problem that they have now depending on jetty-9 and having gwt-dev with jetty-8 inside.

 > * do not have gwt-dev as a dependency

I don't follow here, if they don't depend on gwt-dev they don't depend on current jetty classes inside of it, but thet wont depend on jetty.jar as a transitive dependency of the proposed  gwt-dev

 > * exclude Jetty from the classpath when using gwt-dev (e.g. using
 > runClasspathExcludes with gwt-maven-plugin; that's assuming they
 > don't have GWTTestCase)

But in this case they have more freedom to try to use it's own jetty version (in case it works) excluding the dependency from gwt-dev. Or configure runClasspathExcludes with their jetty version.

 > still that would prevent them from upgrading, and that's not a good
 > thing.

Any upgrade requires some effort, in fact 2.7 already has a bunch of breaking changes.

As reference, Vaadin bundles  most of the gwt-dev stuff in vaadin-client-compiler but it removes some 3party classes which sets as maven dependencies, this is the case of jetty.

 > 
 > Would you also add gwt-codeserver as a dependency?

Do you mean gwt-dev pom including the gwt-codeserver dependency?

currently I'm setting gwt-codeserver as dependency in my projects

 > 
 > I don't think that was the reason, but I just removed the warning.

thanks.

Patch-set: 2