Update patch set 3
Patch Set 3:

re: not overwriting nocache.js

It would be nice to be able to switch back and forth between the production and Super Dev Mode version of the app when we have both versions.  We would need some kind if UI to choose whether to use Super Dev Mode or not. In old DevMode it was triggered by a query parameter. The Super Dev Mode, the bookmarklets set a session variable. Anything new we come up with needs to be at least as secure.

One option might be to save the old version of nocache.js and have a way to put it back somehow.

I think for now, we can tell people to use the bookmarklets and not -launcherDir (or DevMode -superDevMode) if that's what you want.

Patch-set: 3