Call gwt:onLoadErrorFn when loading a permutation fails.
Previously the onLoadErrorFn was only called from a try/catch block
inside of the permutation itself, when it calls the module entry
points (see ModuleUtils.gwtOnLoad).

However, if the permutation never loads in the first place (e.g. due
to a 404), then the user is left with a blank screen, and the
application has no hook to try and recover.

This patch uses the script.onerror property to detect when the browser
fails to load a permutation file and calls the onLoadErrorFn, if
defined.

Change-Id: I9b78e54344fdbe3435d4600a34e4c46e754ee950