Add GSS support.
Second patch concerning the support of GSS in GWT. This patch
introduce the generator that will parse and compile a GSS file
to a css string and generate the implementation of the
corresponding CssResource interface. GSS support is disabled
by default. To enable it, you have to set the configuration
property "CssResource.enableGss" to true.

Change-Id: Iae743842c5707c0231b1c4eb290f7914b6f7a501