Always use generated recompile.nocache.js
With precompile turned off, we were generating the nocache.js with the
same content as a recompile.nocache.js file. This caused issues for
some internal clients loading the application, not being able to
determine the URL of the codeserver.
This patch now always uses stub.nocache.js to generate a nocache.js
file that will include recompile.nocache.js

Change-Id: I17ab3f5691364ea818c7f3db16a4c0eb9aee219e