Starts ignoring modified but unreachable types in incremental.
Previously (in incremental compile mode) unreachable types were not
traversed during the first compile, but if an unreachable type was
edited it would be forced into the staleTypes list, thus causing it to
be traversed.

If the type in question was not GWT compatible the result would be a
compile error that could only be resolved by restarting SuperDevMode.

With this change care is now take to remove any known-unreachable types
from calculated staleTypes lists, so as not to unnecessarily force them
to be traversed.

A new test is added specifically for this behavior.

Sometimes a Generator will create a brand new type and this type while
seen as initially "modified" will be removed from the stale list since
it was not previously reachable. This newly reachable type will still be
traversed since when it is seen by UnifyAst it won't yet have JS. But it
won't show up in the calculated staleTypes list. But it will show up in
the list of processedStaleTypes and tests have been updated to assert
against this more accurate list.

A couple of tests stop generating Foo and start generating Bar and those
tests will no longer see Foo as stale (since it no longer exists to be
processed, thus is not in the processedStaleTypes list). Also one test
was making modifications in an unreachable type and asserting that no
Generators were run. While this behavior previously proved something it
no longer proves anything, so the test was updated to make the type
reachable.

Change-Id: I9b5ddee4a7eec3777e238f30358c19ae3a7f2045