Update patch set 2
Patch Set 2:

(3 comments)

I'm a bit wary of people opening their servers to attack without thinking.

Maybe this should be called "gwt.codeserver.unsafe.host"? This would get people to be a bit more curious and perhaps do a few searches rather than blindly copying it. 

However, that's just a band-aid since they (we) will do it anyway if it's the easiest way to solve the problem, just like with CodeServer's -bindAddress option. Maybe we should explore alternatives?

One possibility would be to add a -policyDir option (similar to the new -launcherDir option) where Super Dev Mode writes policy files. If you share a filesystem with the production server, this would be a more secure way to get the policy files to the right place. Would that work for your use cases?

Patch-set: 2