Make make sure regexps in JsSourceUtils don't cause stack overflow.
Java regular expressions are not compiled into automata but into
a nested matcher structure that is executed in a recursive descent
manner. In particular combinations of alternation (|) and (*)
iteration might result in stack size that depend on the input.

Independent, non capturing, groups ((?>X)) must be used for groups
that are not capturing and "*?" instead of "*" for repetition to
avoid this problem.

Change-Id: Ice0e1cce1a4a284381cfa747dc6ead8ba8b8d324