Update patch set 3
Patch Set 3:

> Not sure if this is the right place to bring this up (-contrib
 > perhaps?), but what happened here, and is this coming back in
 > another way? Without this, there is no clean way for GWT apps to
 > shut down and release all memory being used without the page itself
 > ending and navigating away.

This was basically a failed experiment. There was a lot missing cases and no tests so we decided to abandon it a while but nobody cleaned up the code.

This is a quite difficult job, especially for a framework at the size of GWT and also as soon as you step out of gwt-user it was immediately broken as any third-party can hold a reference. The reason it was implemented was to speed tab switching with multiple apps in a giant app which didn't work out. It is much more saner to keep using frames.

We don't have a plan to bring this back at the moment.

Patch-set: 3
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0