Update patch set 3
Patch Set 3:

My interest is for a team I worked with recently that started reinventing this feature before we noticed that it actually already existed. They then adapted their own code into Disposable and have determined that the system is working well for them (same basic case as you described, multiple sub-apps within a greater mega-app). At least so far.

At least as far as I can tell, library/application code only has a problem if it assigns a 'java' object to a JS property somewhere (without a Disposable to clean up), but aside from that all statics, all instances, etc should be correctly mopped up with the script and the iframe are yanked out of the dom.

Frames are their next bet, and I'll communicate this to them. It is possible that the strictness of frames could be a benefit here, but taking away the ability for any one sub-app to reason about the whole page (sizing/positioning for dialogs/info popups, tooltips, and modal masks are the big ones, at least without reinventing each and every one of those features in a custom way) is a big downside.

If it was experimental, shouldn't it have been Javadoc'd previously as such? Is there a reason we aren't walking the normal deprecation process, or asking if anyone wants to maintain it? I only noticed that this was gone from opening `gwt26_27userApi.conf` for fun to see what had changed. I can't promise that I can take it on (at least not in the hour since I noticed it), or that the users I have in mind will want it for themselves to work on, but would it be worth a shot?

Patch-set: 3