Update patch set 3
Patch Set 3:

> My interest is for a team I worked with recently that started
 > reinventing this feature before we noticed that it actually already
 > existed. They then adapted their own code into Disposable and have
 > determined that the system is working well for them (same basic
 > case as you described, multiple sub-apps within a greater
 > mega-app). At least so far.
 > 
 > At least as far as I can tell, library/application code only has a
 > problem if it assigns a 'java' object to a JS property somewhere
 > (without a Disposable to clean up), but aside from that all
 > statics, all instances, etc should be correctly mopped up with the
 > script and the iframe are yanked out of the dom.
 > 

"If it assigns a 'java' object to a JS property somewhere"

actually this should be

"if 'java' object is reachable from a JS property somewhere"

This makes the problem space much bigger. Every third_party code now
needs to be careful about that. E.g. library code makes an xhr
and has a reference to GWT or the library code listens anything from
the page (local storage, window etc.) and holds a reference to GWT
then this is broken.

 > Frames are their next bet, and I'll communicate this to them. It is
 > possible that the strictness of frames could be a benefit here, but
 > taking away the ability for any one sub-app to reason about the
 > whole page (sizing/positioning for dialogs/info popups, tooltips,
 > and modal masks are the big ones, at least without reinventing each
 > and every one of those features in a custom way) is a big downside.
 > 
 > If it was experimental, shouldn't it have been Javadoc'd previously
 > as such?

Yes, it should have been javadoc'd in the first place that this is
experimental. 

 > Is there a reason we aren't walking the normal deprecation
 > process, or asking if anyone wants to maintain it? I only noticed
 > that this was gone from opening `gwt26_27userApi.conf` for fun to
 > see what had changed. I can't promise that I can take it on (at
 > least not in the hour since I noticed it), or that the users I have
 > in mind will want it for themselves to work on, but would it be
 > worth a shot?

As this was experimental and as it is already broken (GWT.runAsync,
DOM events, ScriptInjector etc are some they I know) and as it wasn't
documented nor advertised anywhere, I just deleted the code before it
gets too late. We are not even sure how much it is broken because we
don't have any tests.

Also I never thought it got considerable usage outside. If that is not
the case and if you feel strongly about it, we can go through the
deprecation process.

In regards to someone else maintaining it:
This is a crosscutting feature, not a satellite piece of code so I rather
not have this complexity in the SDK in the first place because of the
reasons I described earlier.

Patch-set: 3