Update patch set 2
Patch Set 2:

(1 comment)

If they are using DevMode -superDevMode without modifications then we're definitely on the same server (and in the same process, actually). We set gwt.codeserver.port automatcally and setting the host as well would certainly be safe.

I expect that this currently breaks if you pass -bindAddress to DevMode since both servers will be on a different address.

I think you were doing something else like copying files from the war directory of the DevMode server and starting a different server? Do you pick up policy files in a similar way? Maybe we should just arrange for the policy file to be put in the same place that DevMode (without superdevmode) used to do it?

Patch-set: 2