Update patch set 2
Patch Set 2:

This commit is though to work as a discussion about where to maintain this info, it's format etc.

I think it's a good practice to have a README in the root of an opensource project which works as a quick guide about how to build and test it.

Maintaining in tree this kind of information is also easiest to maintain when something changes, like adding a new ant task etc.

Patch-set: 2
Label: Verified=0