Update patch set 2
Patch Set 2: Code-Review-1

> >
 > > As I explained in a comment on issue 8898, I believe this
 > > behavior is by-design.
 > 
 > I think user does not need to know the internals of how a null
 > string is maintained internally as a reference to the json null
 > object, so it makes sense to me to handle it in these methods.

GWT in general doesn't like nulls (and if you ask me, that's a good thing, nulls are harmful, that's why we have Optional).

To me, there's a difference between having

 String str = null;

and

 JsonValue val = null;

In the second case, it's easy to coerce the null into Json.createNull(), they're semantically the same.

A 'null' String is different, the intent of the user is different: he definitely wants a String, so why is there a 'null' in there? In so many cases you want to treat null and the empty string as equivalent values (that's why Guava provides isNullOrEmpty, nullToEmpty and emptyToNull helpers); in this case the intent is not clear, so we'd better just refuse a 'null'.
If 'null' is an acceptable value for you, then just use a JsonValue / JsonString (where the put(String,JsonValue) will do what you want) or the ternary operator:

 obj.put("key", val == null ? Json.createNull() : Json.create(val));

Now the intent is clear and explicit.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1