Update patch set 2
Patch Set 2:

A DTO could also print the JSON to a string in the constructor/setter and parse it in the getter, using a transient variable as a cache if the getter is likely to be called more than once.

This isn't quite the same because Java serialization preserves shared nodes and cycles if the caller chose to build the graph that way, but that seems like a bad idea.

Patch-set: 2