Update patch set 2
Patch Set 2:

It's hard to say which would be faster since standard Java serialization is not terribly fast due to lots of reflection, but then again, neither is scanning text strings for terminators.

But if it's a problem, I think we should work on serialization/deserialization as JSON since that's the common case.

Patch-set: 2