Reimplement Splittable.NULL to support compiler checked-mode
Splittable.NULL in prod mode was a 'null' JS value represented as a JSO,
and relied on GWT not chocking when calling JSO methods on a 'null' value.
This however can break when some optimizations are removed, which is the
case in SuperDevMode. It would also break when checked-mode is turned on
in the compiler (again, this would be the case in SuperDevMode).

Bug: issue 8722
Change-Id: I4d3e655ebe67adf23513d5b8c057cf307bb211a4