Update patch set 3
Patch Set 3:

Can we get this in 2.7-beta1 for testing? And if it breaks too much people and/or tools, then we could revert it for 2.7-final. Any objection?

I'll work on a fix for the gwt-maven-plugin but we already have a clean and simple workaround (that we might even want to be the best practice).

Patch-set: 3