Update patch set 3
Patch Set 3:

I'm happy to take a huge swing here and just get it all, if we think it will be accepted. My goal with this initial touch was to test the waters, see if we think it is even feasible to do this for GWT 2.7 - so far, response seems to be 'no'.

I'll pick this back up again after 2.7 ships (or branches, if I'm feeling ambitious), and will round it out once we're ready to keep moving forward.

Patch-set: 3
Reviewer: Gerrit User 5130 <5130@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0