Fixed API break in AutoDirectionHandler.Target.
As part of the GWT module breakup for modular compilation, the I18n
module was updated to not depend on User. This involved changing
AutoDirectionHandler.Target to no longer depend on HasText. While
the getText() method was added in the initial patch, this left
Target subclasses with an @Override'd setText broken. This patch
allows implementors of Target that are compatible with GWT <=2.6 to
build with 2.7.

See fc1babac93730ec48fe5d56d6c22b3657db9e68a for the original change.

Change-Id: I7f3e62b590153dd3fd4a9de89c2b7b05628bb260