Update patch set 3
Patch Set 3:

(1 comment)

Strange, deploying to a file:// repo is supposed to work (according to docs [1]), and I don't remember having had any issue in the past (though I must confess I haven't run this script in a while).

Could it be somehow due to the version of Maven?

[1] https://maven.apache.org/plugins/maven-deploy-plugin/deploy-file-mojo.html#url

Patch-set: 3
Label: Code-Review=0