Improve resource resolution in a case-insensitive file system.
In case insensitive file systems the Java classloarder, when asked
to find a resource by name, might find one whose name differs from the
one requested just in case. This leads to GWT sometimes finding a
loadign the wrong resource.

This patch add checks to make sure that the resource found by the
classloader is the one requested.

Bug: issue 8716.
Change-Id: I1f427c5ab278dc90e002641f04db6eb90a209cfe