Update patch set 4
Patch Set 4:

(1 comment)

> Also, does this work right with multiple modules? I'm not sure if
 > reload() takes effect right away or if it takes long enough to pop
 > up another dialog. We shouldn't show this dialog again if they
 > already decided to reload.

I would not worry about this, if all modules are in the same codeserver, that I think should almost the cases, if the server is not ready the first module will show the confirm, if the user reloads will try the firs module again, when this first module passes all modules will reach the server not showing the confirm but the compiling dialog. 

If modules come from different servers we should show each dialog of failed modules so as the user knows which servers are down. But they will come sequentially, so the second dialog is not shown until the first one passes, etc.

Normally this dialog will be shown the very first time we load the app, because code server takes some time in setting up the module. 

I have played with this and the user experience is quite good.

Patch-set: 4
Label: Verified=0